main

+++ Helper/app/Http/Controllers/AdminController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Admin;
+use Hash;
 use Illuminate\Http\Request;
+use Session;
 
 class AdminController extends Controller
 {
@@ -14,10 +16,12 @@ class AdminController extends Controller
      */
     public function index()
     {
-       
-        $admin = Admin::all();
-        return view('admin.adminPages.adminInfo')->with('admins', $admin);
-
+        if (Session::has('loginId')) {
+            $admin = Admin::all();
+            return view('admin.adminPages.adminInfo')->with('admins', $admin);
+        } else {
+            return view('admin.adminpages.login');
+        }
     }
 
     /**
@@ -27,9 +31,11 @@ class AdminController extends Controller
      */
     public function create()
     {
-
-        return view('admin.adminPages.create');
-
+        if (Session::has('loginId')) {
+            return view('admin.adminPages.create');
+        } else {
+            return view('admin.adminpages.login');
+        }
     }
 
     /**
@@ -40,24 +46,22 @@ class AdminController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'name' => 'required',
-            'email' => 'required|email|unique:admins',
-            'password' => 'required|min:8',
-        ]);
-        $admin = new Admin;
-        $admin->name = $request->input('name');
-        $admin->email = $request->input('email');
-        $admin->password = $request->input('password');
-         $admin->save();
-        // if($req){
-        //     return back() ->with('sucsess','You have registerd successfuly');
-        // }
-        // else{
-        //     return back() ->with('fail','You have not registerd ');
-        // }
-        // $admin = Admin::create($request->all);
-        return redirect('/admin')->with('success', 'Admin Created');
+        if (Session::has('loginId')) {
+            $this->validate($request, [
+                'name' => 'required',
+                'email' => 'required|email|unique:admins',
+                'password' => 'required|min:8',
+            ]);
+            $admin = new Admin;
+            $admin->name = $request->input('name');
+            $admin->email = $request->input('email');
+            $admin->password = Hash::make($request->input('password'));
+            $admin->save();
+            return redirect('/admin')->with('success', 'Admin Created');
+        } else {
+            return view('admin.adminpages.login');
+        }
+       
 
     }
 
@@ -79,10 +83,12 @@ class AdminController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {
+    {if (Session::has('loginId')) {
         $admin = Admin::find($id);
         return view('admin.adminPages.adminEdit')->with('admins', $admin);
-
+    } else {
+        return view('admin.adminpages.login');
+    }
     }
 
     /**
@@ -94,13 +100,17 @@ class AdminController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $admin = Admin::find($id);
-        $admin->name = $request->input('name');
-        $admin->email = $request->input('email');
-        $admin->password = $request->input('password'); 
-        $admin->save();
-        return redirect('/admin/create')->with('success', "Admin Edited");
+        if (Session::has('loginId')) {
+            $admin = Admin::find($id);
+            $admin->name = $request->input('name');
+            $admin->email = $request->input('email');
+            $admin->password = Hash::make($request->input('password'));
 
+            $admin->save();
+            return redirect('/admin')->with('success', "Admin Edited");
+        } else {
+            return view('admin.adminpages.login');
+        }
     }
 
     /**
@@ -111,27 +121,50 @@ class AdminController extends Controller
      */
     public function destroy($id)
     {
-        $admins = Admin::find($id);
-        $admins->delete();
-        return redirect('/admin')->with('success', 'admin Deleted');
-
+        if (Session::has('loginId')) {
+            $admins = Admin::find($id);
+            $admins->delete();
+            return redirect('/admin')->with('success', 'admin Deleted');
+        } else {
+            return view('admin.adminpages.login');
+        }
     }
 
-    function login(){
+    public function login()
+    {
+        if (Session::has('loginId')) {
         return view('admin.adminpages.login');
+          } else {
+            return view('admin.adminpages.login');
+        }
     }
-    function authLogin(Request $request){
-       $request->validate([
-        'email'=>'required|email|unique:admins',
-        'password'=>'required|min:8'
-       ]);
-       $admin = Admin::where('email', '=', $request->email)->first();
-       if($request->password ===$admin->password){
-            $request->session()->put('loginId',$admin->id);
-            return redirect('/admin');
-       }
-       else{
-        return "not match";
-       }
+    public function authLogin(Request $request)
+    {
+        
+        $request->validate([
+            'email' => 'required|email',
+            'password' => 'required|min:8',
+        ]);
+        $admin = Admin::where('email', '=', $request->email)->first();
+        if ($admin) {
+            if (Hash::check($request->password, $admin->password)) {
+                $name = $request->session()->put('name', $admin->name);
+                $email = $request->session()->put('email', $admin->email);
+
+                $request->session()->put('loginId', $admin->id);
+                return redirect('/admin');
+            } else {
+                return back()->with('fail', 'Password not matches');
+            }
+        } else {
+
+        }
+    }
+
+    public function logout(){
+        if(Session::has('loginId')){
+            Session::pull('loginId');
+            return redirect('login');
+        }
     }
 }
